feat: remove a bunch of warning triggering stuff

+++ common/ntl.c
@@ -155,7 +155,7 @@ ntl_to_buf(char *buf, size_t size, void **p, struct ntl_str_delimiter * d,
   for(i = 0; p[i]; i++) {
     bool is_last = (NULL == p[i+1]);
     psize = (*x)(buf, size, p[i]);
-    if (psize < 0) // error happens
+    if ((long long)psize < 0) // error happens
       return -1;
 
     if(start) {
